Alchemy handler 'bug' fix :P
Just added a break so that the for loop when going through the items when searching uses, exits when it has found the ingredient. This fixes the bug where searching usages on an item that's used up more than once in a recipe results in multiple copies of the recipe.