MineTweaker additions
Harvest Moon gets an add custom crop type that's itemstack sensitive for
crops that share an item for seeds., it's not perfect but it works
pretty well.
Also gets a MT method to interface with this new harvest type
MT functions to add/remove focus items for falling tower as well as
define the meteor size and distribution.