Fix Bound Tool Harvest Levels
This way, they'll be 5 by default, but since they now follow Forge's normal way of getting/setting the harvest level on tools, other mods - such as Iguana Tinker Tweaks - can alter their harvest levels as desired.  When the tools are not activated, they will now return a -1 instead of 0, which results in not being able to be used as a tool at all, rather than the equivalent of a wooden tool.