Setting up workspace now sets a permanent username
This username can be defined in a `gradle.properties` file. For the purposes of this project, use your global property file in `USERHOME\.gradle\gradle.properties`. If no username is set there, it will default to `BloodMagicDev`.

This should help with dev so you don't start with a brand new username each time you load up the game (without manually setting it otherwise)