Started with blocks
Missing entries:
-Blood Letter's Pack
-Rudimentary Snare
-Draft of Angelus
-Tiled Wooden Path
-Tiled Stone Path
-Tiled Worn Stone Path
-Tiled Obsidian Path
-Raw Stone Path
-Small Raw Stone Bricks
-Raw Stone Tiles
-Accented Raw Stone Tiles
-Raw Inversion Pillar
-Raw Inversion Pillar Base
-Raw Inversion Pillar Cap
-Raw Demon Eye

I think raw in this context means pure? I'm not quite sure so I didn't change those yet. Tiled refers to tiles, like in a bathroom? I'd like clarification on that too. Blood Letter I haven't found a translation for yet. The act of bloodletting I know, but not the profession. As for draft of Angelus, is draft there a draft of wind or draft as in a rough draft?

Regarding the sigil stuff: the translations I can think of are the equivalent of sigil, symbol and sign. Sigil would obviously be more precise but the word is more old fashioned and out of use. I don't know if you have a particular preference either way but I prefer it as it is right now.