Continuation of English Translation
Missing entries:
-Blood Letter's Pack
-Rudimentary Snare
-Draft of Angelus
-Tiled Wooden Path
-Tiled Stone Path
-Tiled Worn Stone Path
-Tiled Obsidian Path
-Raw Stone Path
-Small Raw Stone Bricks
-Raw Stone Tiles
-Accented Raw Stone Tiles
-Raw Inversion Pillar
-Raw Inversion Pillar Base
-Raw Inversion Pillar Cap
-Raw Demon Eye
-Polished Raw Stone
-Raw Demon Alloy
-Raw Stone Pillar
-Accented Raw Stone Pillar
-Raw Stone Pillar Cap
-Raw Stone Brick Wall
-Raw Small Stone Brick Wall
-Corrosive Small Stone Brick Wall
-Destructive Small Stone Brick Wall
-Vengeful Small Stone Brick Wall
-Steadfast Small Stone Brick Wall
-Raw Stone Wall
-Raw Stone Stairs
-Stores raw life essence
-Better than a Swiffer®!
-I am very close to being moved by this


I think raw in this context means pure? I'm not quite sure so I didn't change those yet. Tiled refers to tiles, like in a bathroom? I'd like clarification on that too. Blood Letter I haven't found a translation for yet. The act of bloodletting I know, but not the profession. As for draft of Angelus, is draft there a draft of wind or draft as in a rough draft?

New issues arise: not entirely sure what small in "Raw Small Stone Brick Wall" refers to. No clue at all what a swiffer is. Not very sure if there's a way to translate the joke "I am very close to being moved by this". Don't know if I should try to come up with it or to take one of the double meanings (and if so, which one).

Regarding the sigil stuff: the translations I can think of are the equivalent of sigil, symbol and sign. Sigil would obviously be more precise but the word is more old fashioned and out of use. I don't know if you have a particular preference either way but I prefer it as it is right now.