Replaced the Entity registration code. Before, the code was manually adding entities to the GameRegistry.
Doing this prevents the engine from fully registering the added entities in all the proper places. The result was that the Client was never informed of when these entities were created or what they were doing, as all of that behavior is managed by the EntityRegistry.

Changing to the proper EntityRegistry calls fixes issue #1065.