Mob Sacrifice Array no longer kills bosses or players
This is solving the issue I raised in https://github.com/WayofTime/BloodMagic/issues/1368
Added a single check to line 123 of src/main/java/WayofTime/bloodmagic/alchemyArray/AlchemyArrayEffectMobSacrifice.java.
The check makes sure the effective health of the currently targeted entity is above 0. So bosses and players are no longer set to -1 health when the array finishes crafting.