Fix input item routing filters not properly filtering
ItemStacks with count <= 0 are empty, causing them to be treated as a
stack of air. Many comparisons with empty stacks will directly use air
as the item INSTEAD of the ItemStack's actual item.

Input filtering nodes were using negative stack sizes to get the amount
of items that can be pulled, causing incorrect comparisons.