Improved Functional Array Template
This gives the Functional Array template a 32x32 output render rather than the previous 16x16.

The Alchemy Table's output texture was removed to make room.  That template was tweaked to use the soulforge's output texture (which is/was identical)

The dark box was moved to that position (where there was now room) and enlarged.

The Functional Array template had almost all elements shifted for proper allignment.  The new texture was used.  The 'crafting method' item (Arcane Ash) was shifted under the arrow to make room.  The scale of the rendered output texture was doubled from 0.03125 to 0.0625.