Improved Ritual descriptions (et al) (#1761)
* Updated book to match v3.0.4-9.

did a general quality pass, removed extranious $(item) tags where $(l:) tags were already being used, added a page for the explosive charges, added a description of blood transfer to and from the blood altar, added a 2x soulforge template, removed progress bar (for now)

* added changelog

also added links from the changelog to various relevant pages in the book, and added anchors accordingly.

* Update manual to 3.0.6-11

updated changelog, added page for Well of Suffering, rewrote entry on Explosive Charges, slightly tweaked the Anointments to better describe how they wear off, added a $(blood) tag to the manual (same colour as $(fire) so we can talk about Blood more dramatically. This is BLOOD magic after all! (Note, make sure this works - the word 'blood' in the Well of Suffering should be red.

* Improved Ritual Descriptions & Wills

Added links to a few things, formatted some bits to be prettier, elaborated on Fire Fuse, explained that benefits that improve with more Will in the aura don't care about other aspects,

* fixed overlapping URLs

might be a temporary fix until patchouli solves this bug on their end, but the 'NOT YET IMPLEMENTED' link was previously running-over into the clickable area for 'demon will'. Moving it to a new line should solve this.

* Anointments work on Explosives

... unless they don't, that is... Way, be sure to remove the bit for 'looting' working on charges if you can't crack that problem. (I've got faith in you, so i've left it in for now!)

* misc fixes

Cutting Fluid works with the Ritual of the Crusher, and the Self Sacrifice bonus is +150%, not 150%.