Guide Living Armor Upgrade List: Optional Text Components.
I couldn't figure out how to set the height automatically.

Each of these puts a blank space between the list and the text.  "text5" fits it under a List with 5 levels max, and "text10" fits under a list with 10 levels max.

If you need to add more, each layer of text in Patchouli is 9 pixels.