Book: Move Patchouli Formatting Codes to Lang File
This reworks RitualInfoProcessor and en_us.json (and generator) to move the Patchouli formatting codes into the editable language files, rather than being hard-coded in the code.

The color codes for the Rune Types are still hard coded in the Enum, but they can be edited in book.json already.