Add Processor Documentation.  Add Extra Text Capability
Add a comment near the top of all Patchouli Processors documenting their usage.

Added the ability for the RitualInfoProcessor to add extra text specified by the entry below the auto-generated text.

Updated the LivingArmourUpgradeInfoTable to also put entry added text under the generated text.  Removed the hacky (and unused) "text5" and "text10" from the Template.

Slight tweak to RitualInfoProcessor to use a StringBuilder (rather than +=) for the runeCounts variable.