Created ritual_data Template.  Updated RitualInfoProcessor
Created a Ritual Data Template, and Expanded Ritual Info Processor to accommodate it.

The new template recreates Patchouili's Spotlight page, and automatically fills in Will and Range information.

The new page also works with the Text Override feature.  It also includes Overrides for the Heading and Item provided

Usage is documented near the top of the processor.