Polishing up RitualInfoProcessor
Removed unnecessary heading initialization.

Move the extraTest set to the bottom of setup for readability (extra text is the last part of the page).

Removed unnecessary check for pageType being null (it will always be initialized).

removed a Switch with a single If Else statement.

Added a language file localization for a double new line `"$(br2)%s"