Corrected Ritual Info Page Spacing for new Patchouli Version
In the dev version of Patchouli we had been using, `$(br)$(br)` would create a single new line (effectively ignoring one of them).  That is what was referenced when writing the Processor.

Newer versions of Patchouli will cause `$(br)$(br)` to do a double new line.  This update is to account for that change.  Fortunately, all of this formatting is handled in the language files.