Updated Soul Network to second person (#1773)
* Updated book to match v3.0.4-9.

did a general quality pass, removed extranious $(item) tags where $(l:) tags were already being used, added a page for the explosive charges, added a description of blood transfer to and from the blood altar, added a 2x soulforge template, removed progress bar (for now)

* added changelog

also added links from the changelog to various relevant pages in the book, and added anchors accordingly.

* Update manual to 3.0.6-11

updated changelog, added page for Well of Suffering, rewrote entry on Explosive Charges, slightly tweaked the Anointments to better describe how they wear off, added a $(blood) tag to the manual (same colour as $(fire) so we can talk about Blood more dramatically. This is BLOOD magic after all! (Note, make sure this works - the word 'blood' in the Well of Suffering should be red.

* Improved Ritual Descriptions & Wills

Added links to a few things, formatted some bits to be prettier, elaborated on Fire Fuse, explained that benefits that improve with more Will in the aura don't care about other aspects,

* fixed overlapping URLs

might be a temporary fix until patchouli solves this bug on their end, but the 'NOT YET IMPLEMENTED' link was previously running-over into the clickable area for 'demon will'. Moving it to a new line should solve this.

* Anointments work on Explosives

... unless they don't, that is... Way, be sure to remove the bit for 'looting' working on charges if you can't crack that problem. (I've got faith in you, so i've left it in for now!)

* misc fixes

Cutting Fluid works with the Ritual of the Crusher, and the Self Sacrifice bonus is +150%, not 150%.

* added bloodmagic namespace to links

... editing, like, 3/4 of all files. Hoo boy.

* fixed some boo-boos

Whoops. Also a little more polish here and there. The usual.

* Add Bloodmagic's namespace to Category calls

This has to be updated in every single Entry, and the 3 current sub-categories.

* Corrected Soul Snare's recipe reference

It's now crafted in a Blood Altar.

* Updated Charges

added Fungal and Controlled charges, added the Charge Anointments, added a 3x Soulforge Recipe page to handle them.

* added anchors to new entries

* Create throwing_daggers.json

Ouch! Hey, these are pointy!

* added entry for alchemy table's GUI

* Updated 'Getting Started'

to make note of recent changes to progression - first gold, then the altar, then the snares. Tweaked several related pages accordingly. I think i caught them all!

* Update getting_started.json

changed it to second person to match the rest of the book and added a note about rituals causing nausea. (The note about LP-using items hurting you was already in there, it appears)

* Update soul_network.json

...well bum, my description of the last change was meant for this one, and vice versa. Oh well; in the other one i added some links and recoloured the words Blood Magic to be suitably Blood-coloured.

Co-authored-by: VT-14 <GamingVT14@gmail.com>