Book: Ritual Info Page Processor (#1821)
* Book: Ritual Info Page Processor

This creates a Patchouli Processor which inputs a Ritual ID, and will return that Ritual's info blurb, rune counts, Activation Crystal tier, Activation Cost, and any Upkeep Costs. It uses the Ritual Diviner's language keys wherever possible to avoid duplication.

EnumRuneType was extended to also include a translation suffix, and a Patchouli color code (uses the macro names, actual hex colors are set in book.json).

Added a Ritual Stone counting method (countRunes) to RitualHelper.

Changed ItemRitualDiviner to use the new RitualHelper.countRunes method.

* Book: Improved RitualHelper and Cleaned Up EnumRuneType

Cleaned up EnumRitualType so the translation key and patchouli color use the value's name (forcing consistency).

Significantly improved the counting method in RitualHelper.

Thanks to TehNut on Discord for the guidance.

* Book: Move Patchouli Formatting Codes to Lang File

This reworks RitualInfoProcessor and en_us.json (and generator) to move the Patchouli formatting codes into the editable language files, rather than being hard-coded in the code.

The color codes for the Rune Types are still hard coded in the Enum, but they can be edited in book.json already.

* deduplicated data from en_us.json. added a new page about upgrade tomes. Slightly edited some of the Ritual Diviner's lines to match what was previously in the manual.

* Put Wrince's language changes in the Language Generator

...so they don't get overwritten immediately. :stuck_out_tongue:

Co-Authored-By: wrincewind <1457878+wrincewind@users.noreply.github.com>

* Add Processor Documentation.  Add Extra Text Capability

Add a comment near the top of all Patchouli Processors documenting their usage.

Added the ability for the RitualInfoProcessor to add extra text specified by the entry below the auto-generated text.

Updated the LivingArmourUpgradeInfoTable to also put entry added text under the generated text.  Removed the hacky (and unused) "text5" and "text10" from the Template.

Slight tweak to RitualInfoProcessor to use a StringBuilder (rather than +=) for the runeCounts variable.

* Add a formatting code override to RitualInfoProcessor

This allows us to add Patchouli Coded (color, links, etc.) to the generated text coming from the Ritual Diviner entries.

Applied to the ritual_crystal_harvest entry as an example/test.

* Apply RitualInfoProcessor Text Overrides to Documented Rituals

Just adding links and coloring to match the rest of the book.

* Created ritual_data Template.  Updated RitualInfoProcessor

Created a Ritual Data Template, and Expanded Ritual Info Processor to accommodate it.

The new template recreates Patchouili's Spotlight page, and automatically fills in Will and Range information.

The new page also works with the Text Override feature.  It also includes Overrides for the Heading and Item provided

Usage is documented near the top of the processor.

* Polishing up RitualInfoProcessor

Removed unnecessary heading initialization.

Move the extraTest set to the bottom of setup for readability (extra text is the last part of the page).

Removed unnecessary check for pageType being null (it will always be initialized).

removed a Switch with a single If Else statement.

Added a language file localization for a double new line `"$(br2)%s"

* Book: Add Dusk and Dawn color codes.  Fix Ritual List text link

* Use empty List and DemonWillHolder for default range info

An empty list of EnumDemonWillType = no Will Types used.  Null would make certain ritual processors crash on that code.

I also gave it a new (presumably empty) DemonWillHolder, in case that ever becomes an issue.

* Better Empty List for RitualInfoProcessor

* updated documentation on will-types and their uses in rituals

* added colours for dusk and dawn runes

* made rituals more dynamic
now they pull their data from the language files. whee!

* VT's Ritual Entry Tweaks

A handful of content/formatting tweaks for the Book's Ritual Entries.

* Add Curios Support

Add Curios Support for Tartaric Gems (Necklace) , Sigils (Charm), and the Tome of Peritia (Charm).

* Add Curios API to the Dev Environment.
* Item Tag Tartaric Gems as a Curios Necklace, and relevant Sigils and Tome of Peritia as Curios Charms.
* Load CuriosCompat from BloodMagic.java if Curios is installed.
* Adjust the Blood Magic API to register an inventory provider.
* CuriosCompat uses an IMC to request at least 1 Necklace and 1 Charm slot to be available to the player, and registers the Curios Inventory with the API.
* Made an InventoryHelper class.  Right now it only has one method which returns all available API inventory items.
* Adjusted PlayerDemonWillHandler and Generic Handler (Tome of Peritia's XP event listener) to use the new InventoryHelper.

* Add Blocked Build Error Message to Ritual Diviner

"Unable to replace block at X, Y, Z."

* Corrected ritual.notValid translation usage.

* Revert "Merge branch 'VT_1.16.3' into Book-Ritual_Processor"

This reverts commit 1947486e5e01434d279cf1b6e6c996faa8ffba8d, reversing
changes made to b1834f29230cbb376d03c51167702f36c4afcacd.

* Upgrade Build Environment's Patchouli Version

Going from a 48-Snapshot build to the release version 53 build.

* Corrected Ritual Info Page Spacing for new Patchouli Version

In the dev version of Patchouli we had been using, `$(br)$(br)` would create a single new line (effectively ignoring one of them).  That is what was referenced when writing the Processor.

Newer versions of Patchouli will cause `$(br)$(br)` to do a double new line.  This update is to account for that change.  Fortunately, all of this formatting is handled in the language files.

* Re-add the "weak" Link Anchor for Weak Activation Crystal.

This got dropped in a conflict correction merge.

Co-authored-by: wrincewind <Cott002367@yahoo.co.uk>
Co-authored-by: wrincewind <1457878+wrincewind@users.noreply.github.com>