Demon Will Aura Gauge recipe, lang, & documentation
The Demon Will was already functional, but missing a recipe and language files.  I re-created what it previously had for those.

I also started documenting it in the Sanguine Scientiem.  It has an entry that describes what it does, but to really work we need to refactor part of the book to address what the Aura actually is.  *pokes Wrince*.