Book: Register Patchouli Multiblocks (#1826)
Added a RegisterPatchouliMultiblocks class.  This takes in all of the Rituals and Blood Altar multiblocks (defined by a list of components) and converts them to the form that Patchouli understands.

This change makes updating the book significantly easier as we just need to supply the corresponding Multiblock ID (`"bloodmagic:ritualname"`) rather than manually making ASCII Art for the multiblock.

In addition, this change gives us better control over what is displayed and works for multiblock verification, especially for the Blood Altar.
* The Tier-2 corner Runes which don't apply a bonus appear only as Blank Runes, but accept any Blood Rune for verification.
* The "Not Air" spaces (Pillars) use to check for specifically Stone Bricks.  They now display Stone Bricks but accept any block that isn't Air or a Liquid (the same basic check that the Altar itself uses for those positions).
* The ComponentTypes for BloodRunes, Glowstone, BloodStone, Beacons, and Crystals should pull from the API values for those types, improving add-on compatibility and eliminating the need for several messy Block Tags.

All Rituals are submitted using the North orientation.  Patchouli automatically rotates it depending on which direction the player is facing when the render is anchored.

****

Related Minor Changes:
* The "WayofTime-Dev" formatter was edited to allow the usage of Formatter Off/On codes.  This was used in two places to keep that part of the code readable.
* During `OnLoadComplete`, BloodMagic.java checks to see if Patchouli is in the mod list, and triggers RegisterPatchouliMultiblocks if it is.  I tested a build without Patchouli to confirm that it doesn't crash without Patchouli.