Give Recipe, Lang, and Documentation to Demon Will Aura Gauge.  Tweak Crystal Harvest. (#1829)
* Demon Will Aura Gauge recipe, lang, & documentation

The Demon Will was already functional, but missing a recipe and language files.  I re-created what it previously had for those.

I also started documenting it in the Sanguine Scientiem.  It has an entry that describes what it does, but to really work we need to refactor part of the book to address what the Aura actually is.  *pokes Wrince*.

* Added info on the Aura, tidied up related pages.

* Book: Aura Related Tweaks

* Let Will Crystal Harvest use any Will Type

Ha ha Auto-formatter go crunch.

Also includes some more book tweaks.

This is the "any of 1 type of will" version.

Co-authored-by: wrincewind <Cott002367@yahoo.co.uk>