Implemented functionality to the Master Routing Node
Also includes an updated GUI for the input and output routing nodes.
Changed the `ItemStack` filters to a new `IFilterKey` when checking if an item is valid for a given filter. This fixes the MRN having issues with negative item stack sizes, while also allowing for a better way to check filters.