Fixed the trainer's logic and fixed an edge case for the Elytra upgrade rendering.
Trainer will now allow exp up to the specified level if it is a new level, and will prevent subsequent exp from being gained. For instance, if the Trainer is set to Tough 4, and you are at Tough 3, it will allow you to level up to Tough 4, but will not increase in exp towards Tough 5.

Also, fixes an edge case for the elytra rendering for creative-spawned chestpieces.