Added Gilded Living Armour Upgrade
Gilded makes the Piglins found across the Nether passive, as if you were wearing a piece of Gold Armor.

Without this upgrade, Piglins are hostile, and appeasing them by wearing a piece of Gold Armor means you aren't wearing a full set of Living Armour, so you don't get the buffs.

Current Stats: Costs 5 Upgrade Points.  Crafted using a Binding Array and a Block of Gold.

This includes documentation in the Sanguine Scientiem.