Implement the Gilded Upgrade
The upgrade works, but is only obtainable by spawning it in (Commands or the Creative Tab).

Still need to work out a way to Train the upgrade.  I'm planning on a certain number of Piglin Trades, but am not sure how to properly track that (can't find an event for it).