Add a ModID Check to Meteor Recipes
Adds a ModID Check to Meteor Recipes.  The "modid" field is optional and will use our ID if not specified, which is obviously loaded.

Not loaded ModIDs return null, which skips that recipe.  This is a single line in the log rather than a large stacktrace.

Added ModIDs (copied from the 1.16 guidebook documentation) to the 5 mod-compat meteors.