Expand InventoryHelper & Update PlayerUtil.findItem().
Expand InventoryHelper with methods to find a specific inventory, and all registered inventories excluding certain ones.  Together these allow a method to check inventories in a specific order, but still be able to check all registered inventories.

Update PlayerUtil.findItem() to use the update InventoryHelper.  This is needed to allow it to check Curios and any add-on mod inventories.  Decided to check Curios before the Main Inventory, but after the Off-Hand slot.